tag object commands simplified
