+++ tags/__init__.py
@@ -28,7 +28,7 @@ from pathlib import Path
 from redbot.core.bot import Red
 from redbot.core.errors import CogLoadError
 
-from .tags import Tags
+from .core import Tags
 
 with open(Path(__file__).parent / "info.json") as fp:
     __red_end_user_data_statement__ = json.load(fp)["end_user_data_statement"]
